{{cookiecutter.project_name}}/requirements-dev.txt
gcloud-utils>=2.17.9
nose>=1.3.7
pylint>=1.9.4
twine>=1.13.0